Move utility functions into pydot.utils
This way they can be shared between core and dot_parser.
