+++ src/pydot/core.py
@@ -9,7 +9,6 @@ import errno
 import itertools
 import logging
 import os
-import re
 import subprocess
 import sys
 import warnings
@@ -21,11 +20,16 @@ from pydot.classes import AttributeDict, FrozenDict
 from pydot.constants import (
     CLUSTER_ATTRIBUTES,
     DEFAULT_PROGRAMS,
-    DOT_KEYWORDS,
     EDGE_ATTRIBUTES,
     GRAPH_ATTRIBUTES,
     NODE_ATTRIBUTES,
     OUTPUT_FORMATS,
+)
+from pydot.utils import (
+    format_for_lookup,
+    possibly_unquoted,
+    quote_attr_if_necessary,
+    quote_id_if_necessary,
 )
 
 _logger = logging.getLogger(__name__)
@@ -160,135 +164,6 @@ def call_graphviz(
 
     return stdout_data, stderr_data, process
 
-
-def make_quoted(s: str) -> str:
-    """Transform a string into a quoted string, escaping internal quotes."""
-    out = s.replace('"', r"\"")
-    return f'"{out}"'
-
-
-re_numeric = re.compile(r"^-?([0-9]+\.?[0-9]*|[0-9]*\.[0-9]+)$")
-re_dbl_quoted = re.compile(r'^".*"$', re.S)
-re_html = re.compile(r"^<.*>$", re.S)
-
-id_re_alpha_nums = re.compile(r"^[_a-zA-Z][a-zA-Z0-9_]*$")
-id_re_alpha_nums_with_ports = re.compile(
-    r'^[_a-zA-Z][a-zA-Z0-9_:"]*[a-zA-Z0-9_"]+$'
-)
-id_re_with_port = re.compile(r"^([^:]*):([^:]*)$")
-
-
-def any_needs_quotes(s: str) -> Optional[bool]:
-    """Determine if a string needs to be quoted.
-
-    Returns True, False, or None if the result is indeterminate.
-    """
-
-    # Strings consisting _only_ of digits are safe unquoted
-    if s.isdigit():
-        return False
-
-    # MIXED-aphanumeric values need quoting if they *start* with a digit
-    if s.isalnum():
-        return s[0].isdigit()
-
-    for test_re in [re_numeric, re_dbl_quoted, re_html]:
-        if test_re.match(s):
-            return False
-
-    has_high_chars = any(ord(c) > 0x7F or ord(c) == 0 for c in s)
-    if has_high_chars:
-        return True
-
-    return None
-
-
-def id_needs_quotes(s: str) -> bool:
-    """Checks whether a string is a dot language ID.
-
-    It will check whether the string is solely composed
-    by the characters allowed in an ID or not.
-    If the string is one of the reserved keywords it will
-    need quotes too but the user will need to add them
-    manually.
-    """
-
-    # If the name is a reserved keyword it will need quotes but pydot
-    # can't tell when it's being used as a keyword or when it's simply
-    # a name. Hence the user needs to supply the quotes when an element
-    # would use a reserved keyword as name. This function will return
-    # false indicating that a keyword string, if provided as-is, won't
-    # need quotes.
-    if s.lower() in DOT_KEYWORDS:
-        return False
-
-    any_result = any_needs_quotes(s)
-    if any_result is not None:
-        return any_result
-
-    for test_re in [
-        id_re_alpha_nums,
-        id_re_alpha_nums_with_ports,
-    ]:
-        if test_re.match(s):
-            return False
-
-    m = id_re_with_port.match(s)
-    if m:
-        return id_needs_quotes(m.group(1)) or id_needs_quotes(m.group(2))
-
-    return True
-
-
-def quote_id_if_necessary(
-    s: str, unquoted_keywords: Optional[Sequence[str]] = None
-) -> str:
-    """Enclose identifier in quotes, if needed."""
-    unquoted = [
-        w.lower() for w in list(unquoted_keywords if unquoted_keywords else [])
-    ]
-
-    if isinstance(s, bool):
-        return str(s).lower()
-    if not isinstance(s, str):
-        return s
-    if not s:
-        return s
-
-    if s.lower() in unquoted:
-        return s
-    if s.lower() in DOT_KEYWORDS:
-        return make_quoted(s)
-
-    if id_needs_quotes(s):
-        return make_quoted(s)
-
-    return s
-
-
-def quote_attr_if_necessary(s: str) -> str:
-    """Enclose attribute value in quotes, if needed."""
-    if isinstance(s, bool):
-        return str(s).lower()
-
-    if not isinstance(s, str):
-        return s
-
-    if s.lower() in DOT_KEYWORDS:
-        return make_quoted(s)
-
-    any_result = any_needs_quotes(s)
-    if any_result is not None and not any_result:
-        return s
-
-    return make_quoted(s)
-
-
-def format_for_lookup(s: str) -> List[str]:
-    """Return a list of the possible lookup forms of an identifier."""
-    if re_dbl_quoted.match(s) or re_html.match(s):
-        return [s]
-    return [s, f'"{s}"']
 
 
 def graph_from_dot_data(s: str) -> Optional[List["Dot"]]:
